// institutions/NeispuoRegInstFront/src/app/enums/menu.enum.ts
export enum Menu {
  Home = "home",
  Active = "active-institutions",
  Inactive = "inactive-institutions",
  Checkup = "check-up",
  Settings = "settings",
  NewInstitution = "new-institution",
  CreateProcedure = "create-procedure",
  EditProcedure = "edit-procedure",
  Unauthorized = "unauthorized"
}

export enum MenuInt {
  "home",
  "active-institutions",
  "inactive-institutions",
  "check-up",
  "settings",
  "new-institution",
  "create-procedure"
}