// institutions/NeispuoRegInstFront/src/app/enums/influenceType.enum.ts
export enum InfluenceType {
  Options = "options",
  Disable = "disable",
  Hide = "hide",
  HideClear = "hideAndClear",
  SetValueOppositeCondition = "setValueOppositeCondition",
  SetValue = "setValue",
  DefaultValue = "defaultValue", // should be reset on influencer reset if no db value provided
  Render = "render", //opposite of hide 
  Require = "require"
}